refactor navbar, mobile nav modal
