+++ src/components/navbar/nav_links.js
@@ -1,0 +1,23 @@
+export default [
+  {
+    id: "nav.our_approach",
+    defMsg: "Our Approach",
+    path: "/our-approach",
+  },
+  {
+    id: "nav.our_org",
+    defMsg: "Our Org",
+    path: "/team",
+  },
+  {
+    id: "nav.in_the_news",
+    defMsg: "In the News",
+    path: "/news",
+  },
+  {
+    id: "nav.get_involved",
+    defMsg: "GET INVOLVED",
+    path: "/get-involved",
+    classes: "text-white bg-primary py-2 px-6",
+  },
+];
